+++ clients/ctl/admin-ui/src/pages/system/new/index.tsx
@@ -5,7 +5,6 @@ import {
   Button,
   Heading,
   Stack,
-  Text,
 } from "@fidesui/react";
 import type { NextPage } from "next";
 import NextLink from "next/link";
@@ -15,9 +14,8 @@ import Layout from "~/features/common/Layout";
 import SystemYamlForm from "~/features/system/SystemYamlForm";
 
 const NewSystem: NextPage = () => {
-  const [generateMethod, setGenerateMethod] = useState<
-    "yaml" | "manual" | null
-  >(null);
+  const [showYamlForm, setShowYamlForm] = useState(false);
+
   return (
     <Layout title="Systems">
       <Heading mb={2} fontSize="2xl" fontWeight="semibold">
@@ -34,29 +32,35 @@ const NewSystem: NextPage = () => {
         </Breadcrumb>
       </Box>
       <Stack spacing={8}>
-        <Box w={{ base: "100%", lg: "50%" }}>
+        {/* <Box w={{ base: "100%", lg: "50%" }}>
           <Text>
             Choose whether to upload a new system YAML or manually generate a
             system.
           </Text>
-        </Box>
+        </Box> */}
         <Box>
           <Button
             size="sm"
             mr={2}
             variant="outline"
-            onClick={() => setGenerateMethod("yaml")}
-            isActive={generateMethod === "yaml"}
+            onClick={() => setShowYamlForm(true)}
+            isActive={showYamlForm}
             data-testid="upload-yaml-btn"
           >
             Upload a new system YAML
           </Button>
-          <Button size="sm" variant="outline">
-            Manually generate a system
-          </Button>
+          {/* <Button
+            size="sm"
+            variant="outline"
+            data-testid="manually-generate-btn"
+          >
+            <NextLink href="/system/new/configure">
+              Manually generate a system
+            </NextLink>
+          </Button> */}
         </Box>
         <Box w={{ base: "100%", lg: "50%" }}>
-          {generateMethod === "yaml" ? <SystemYamlForm /> : null}
+          {showYamlForm ? <SystemYamlForm /> : null}
         </Box>
       </Stack>
     </Layout>